zackhub/requirements.txt
anyascii==0.1.7
asgiref==3.3.1
autopep8==1.5.5
beautifulsoup4==4.8.2
boto3==1.17.17
botocore==1.20.17
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-extensions==3.1.1
django-filter==2.4.0
django-modelcluster==5.1
django-on-heroku==1.1.2
django-phonenumber-field==5.0.0
django-recaptcha==2.0.6
django-storages==1.11.1
django-taggit==1.3.0
django-treebeard==4.5.1
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
l18n==2020.6.1
openpyxl==3.0.6
phonenumbers==8.12.18
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
Unidecode==1.2.0
urllib3==1.26.3
wagtail==2.12.3
wagtailfontawesome==1.2.1
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
